Split resolver phase2 tests
