+++ tests/resolver_phase2.rs
@@ -10,2212 +10,19 @@ fn parse_program(src: &str) -> zen::ast::Program {
     parser::parse(tokens, file_id).expect("parse")
 }
 
-#[test]
-fn resolver_assigns_symbol_ids_in_separate_namespaces() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-Point = () i32 { 1 }
-Color: Red, Blue
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-    let point_type = table.lookup(Namespace::Type, "Point").expect("Point type");
-    let point_value = table
-        .lookup(Namespace::Value, "Point")
-        .expect("Point function");
-    let red_variant = table
-        .lookup(Namespace::Variant, "Red")
-        .expect("Red variant");
-
-    assert_ne!(point_type.id, point_value.id);
-    assert_ne!(point_type.id, red_variant.id);
-    assert_eq!(point_type.name, "Point");
-    assert_eq!(point_value.name, "Point");
-    assert_eq!(red_variant.name, "Red");
-    assert!(point_type.definition_span.start < point_type.definition_span.end);
-}
-
-#[test]
-fn resolver_rejects_duplicate_names_in_same_namespace() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-Point: { y: i32 }
-"#,
-    );
-
-    let err = Resolver::new()
-        .resolve_program(&program)
-        .expect_err("duplicate type name should fail");
-    assert!(
-        err.iter()
-            .any(|d| d.message.contains("duplicate type symbol 'Point'")),
-        "expected duplicate type diagnostic, got {err:?}"
-    );
-}
-
-#[test]
-fn resolver_records_public_visibility_for_exported_declarations() {
-    let program = parse_program(
-        r#"
-pub PublicPoint: { x: i32 }
-PrivatePoint: { x: i32 }
-pub Json<T>: behavior { encode: (Self) T }
-InternalJson: behavior { encode: (Self) i32 }
-pub exported = () i32 { 1 }
-internal = () i32 { 2 }
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-
-    assert!(
-        table
-            .lookup(Namespace::Type, "PublicPoint")
-            .expect("public type")
-            .is_public
-    );
-    assert!(
-        !table
-            .lookup(Namespace::Type, "PrivatePoint")
-            .expect("private type")
-            .is_public
-    );
-    assert!(
-        table
-            .lookup(Namespace::Behavior, "Json")
-            .expect("public behavior")
-            .is_public
-    );
-    assert!(
-        !table
-            .lookup(Namespace::Behavior, "InternalJson")
-            .expect("private behavior")
-            .is_public
-    );
-    assert!(
-        table
-            .lookup(Namespace::Value, "exported")
-            .expect("public function")
-            .is_public
-    );
-    assert!(
-        !table
-            .lookup(Namespace::Value, "internal")
-            .expect("private function")
-            .is_public
-    );
-}
-
-#[test]
-fn resolver_rejects_unknown_type_references_in_declarations() {
-    let program = parse_program(
-        r#"
-Point: { next: MissingPoint }
-distance = (point: Point) UnknownReturn { 0 }
-"#,
-    );
-
-    let err = Resolver::new()
-        .resolve_program(&program)
-        .expect_err("unknown type references should fail");
-
-    assert!(
-        err.iter()
-            .any(|d| d.message.contains("unknown type symbol 'MissingPoint'")),
-        "expected missing field type diagnostic, got {err:?}"
-    );
-    assert!(
-        err.iter()
-            .any(|d| d.message.contains("unknown type symbol 'UnknownReturn'")),
-        "expected missing return type diagnostic, got {err:?}"
-    );
-}
-
-#[test]
-fn resolver_rejects_method_on_unknown_type() {
-    let program = parse_program(
-        r#"
-Missing.label = () str { "missing" }
-"#,
-    );
-
-    let err = Resolver::new()
-        .resolve_program(&program)
-        .expect_err("method receiver type should be known");
-
-    assert!(
-        err.iter()
-            .any(|d| d.message.contains("unknown type symbol 'Missing'")),
-        "expected unknown method receiver type diagnostic, got {err:?}"
-    );
-}
-
-#[test]
-fn resolver_records_method_signatures_as_value_symbols() {
-    let program = parse_program(
-        r#"
-Box<T>: {
-    value: T
-}
-
-Box.get<T> = (self: Box<T>) T {
-    self.value
-}
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-    let method = table
-        .lookup(Namespace::Value, "Box.get")
-        .expect("method symbol");
-
-    assert_eq!(method.parameter_count, Some(1));
-    assert_eq!(
-        method.parameter_names.as_deref(),
-        Some(&["self".to_string()][..])
-    );
-    assert_eq!(
-        method.parameter_type_names.as_deref(),
-        Some(&["Box<T>".to_string()][..])
-    );
-    assert_eq!(method.return_type_name.as_deref(), Some("T"));
-    assert_eq!(method.type_parameter_count, Some(1));
-    assert_eq!(
-        method.type_parameter_names.as_deref(),
-        Some(&["T".to_string()][..])
-    );
-    assert_eq!(method.type_parameter_bounds.as_deref(), Some(&[][..]));
-}
-
-#[test]
-fn resolver_records_method_function_type_signatures() {
-    let program = parse_program(
-        r#"
-Box<T>: {
-    value: T
-}
-
-Box.map<T> = (self: Box<T>, callback: (T) T) (T) T {
-    callback
-}
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-    let method = table
-        .lookup(Namespace::Value, "Box.map")
-        .expect("method symbol");
-
-    assert_eq!(method.parameter_count, Some(2));
-    assert_eq!(
-        method.parameter_names.as_deref(),
-        Some(&["self".to_string(), "callback".to_string()][..])
-    );
-    assert_eq!(
-        method.parameter_type_names.as_deref(),
-        Some(&["Box<T>".to_string(), "(T) T".to_string()][..])
-    );
-    assert_eq!(method.return_type_name.as_deref(), Some("(T) T"));
-    assert_eq!(method.type_parameter_count, Some(1));
-    assert_eq!(
-        method.type_parameter_names.as_deref(),
-        Some(&["T".to_string()][..])
-    );
-    assert_eq!(method.type_parameter_bounds.as_deref(), Some(&[][..]));
-}
-
-#[test]
-fn resolver_rejects_self_type_outside_method_or_behavior() {
-    let program = parse_program(
-        r#"
-main = (value: Self) i32 { 0 }
-"#,
-    );
-
-    let err = Resolver::new()
-        .resolve_program(&program)
-        .expect_err("Self should require a method or behavior context");
-
-    assert!(
-        err.iter()
-            .any(|d| d.message.contains("Self type is only valid")),
-        "expected invalid Self type diagnostic, got {err:?}"
-    );
-}
-
-#[test]
-fn resolver_records_import_bindings_as_symbols() {
-    let program = parse_program(
-        r#"
-{ ExternalPoint, helper } = geometry
-distance = (point: ExternalPoint) i32 { helper() }
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-    let module = table
-        .lookup(Namespace::Module, "geometry")
-        .expect("module symbol");
-    let imported_type = table
-        .lookup(Namespace::Import, "ExternalPoint")
-        .expect("imported type binding");
-    let imported_value = table
-        .lookup(Namespace::Import, "helper")
-        .expect("imported value binding");
-
-    assert_ne!(module.id, imported_type.id);
-    assert_ne!(imported_type.id, imported_value.id);
-    assert_eq!(imported_type.import_source.as_deref(), Some("geometry"));
-    assert_eq!(imported_value.import_source.as_deref(), Some("geometry"));
-}
-
-#[test]
-fn resolver_records_behavior_impl_methods_as_value_symbols() {
-    let program = parse_program(
-        r#"
-Json: behavior {
-    stringify: (Self) str
-}
-
-Point: { x: i32 }
-
-Point.implements(Json) {
-    stringify = (value: Point) str { "point" }
-}
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-    let method = table
-        .lookup(Namespace::Value, "Point.stringify")
-        .expect("impl method symbol");
-
-    assert_eq!(method.name, "Point.stringify");
-    assert_eq!(method.parameter_count, Some(1));
-    assert_eq!(
-        method.parameter_names.as_deref(),
-        Some(&["value".to_string()][..])
-    );
-    assert_eq!(
-        method.parameter_type_names.as_deref(),
-        Some(&["Point".to_string()][..])
-    );
-    assert_eq!(method.return_type_name.as_deref(), Some("str"));
-    assert_eq!(method.type_parameter_count, Some(0));
-    assert_eq!(method.type_parameter_names.as_deref(), Some(&[][..]));
-    assert_eq!(method.type_parameter_bounds.as_deref(), Some(&[][..]));
-}
-
-#[test]
-fn resolver_rejects_duplicate_behavior_impl_edges() {
-    let program = parse_program(
-        r#"
-Marker: behavior { }
-
-Point: { x: i32 }
-
-Point.implements(Marker) { }
-
-Point.implements(Marker) { }
-"#,
-    );
-
-    let err = Resolver::new()
-        .resolve_program(&program)
-        .expect_err("duplicate behavior impl should fail in resolver");
-
-    assert!(
-        err.iter().any(|d| d
-            .message
-            .contains("duplicate behavior implementation `Marker`")),
-        "expected duplicate behavior implementation diagnostic, got {err:?}"
-    );
-}
-
-#[test]
-fn resolver_accepts_non_behavior_impl_blocks_as_method_symbols() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Point.impl = {
-    get = (self: Point) i32 { self.x }
-}
-"#,
-    );
-
-    let symbols = Resolver::new()
-        .resolve_program(&program)
-        .expect("non-behavior impl blocks should resolve");
-
-    let get = symbols
-        .lookup(Namespace::Value, "Point.get")
-        .expect("impl method symbol");
-    assert_eq!(get.parameter_count, Some(1));
-    assert_eq!(
-        get.parameter_type_names.as_ref().map(Vec::as_slice),
-        Some(&["Point".to_string()][..])
-    );
-}
-
-#[test]
-fn resolver_rejects_duplicate_non_behavior_impl_method_names() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Point.impl = {
-    get = (self: Point) i32 { self.x }
-    get = (self: Point) i32 { self.x }
-}
-"#,
-    );
-
-    let err = Resolver::new()
-        .resolve_program(&program)
-        .expect_err("duplicate non-behavior impl methods should fail in resolver");
-
-    assert!(
-        err.iter()
-            .any(|d| d.message.contains("duplicate value symbol 'Point.get'")),
-        "expected duplicate impl method symbol diagnostic, got {err:?}"
-    );
-}
-
-#[test]
-fn resolver_rejects_non_behavior_impl_method_colliding_with_top_level_method() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Point.get = (self: Point) i32 { self.x }
-
-Point.impl = {
-    get = (self: Point) i32 { self.x }
-}
-"#,
-    );
-
-    let err = Resolver::new()
-        .resolve_program(&program)
-        .expect_err("impl method colliding with top-level method should fail in resolver");
-
-    assert!(
-        err.iter()
-            .any(|d| d.message.contains("duplicate value symbol 'Point.get'")),
-        "expected duplicate method symbol diagnostic, got {err:?}"
-    );
-}
-
-#[test]
-fn resolver_records_behavior_impl_function_type_methods() {
-    let program = parse_program(
-        r#"
-Mapper: behavior {
-    map: (Self, (i32) i32) (i32) i32
-}
-
-Point: { x: i32 }
-
-Point.implements(Mapper) {
-    map = (value: Point, callback: (i32) i32) (i32) i32 {
-        callback
-    }
-}
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-    let method = table
-        .lookup(Namespace::Value, "Point.map")
-        .expect("impl method symbol");
-
-    assert_eq!(method.name, "Point.map");
-    assert_eq!(method.parameter_count, Some(2));
-    assert_eq!(
-        method.parameter_names.as_deref(),
-        Some(&["value".to_string(), "callback".to_string()][..])
-    );
-    assert_eq!(
-        method.parameter_type_names.as_deref(),
-        Some(&["Point".to_string(), "(i32) i32".to_string()][..])
-    );
-    assert_eq!(method.return_type_name.as_deref(), Some("(i32) i32"));
-    assert_eq!(method.type_parameter_count, Some(0));
-    assert_eq!(method.type_parameter_names.as_deref(), Some(&[][..]));
-    assert_eq!(method.type_parameter_bounds.as_deref(), Some(&[][..]));
-}
-
-#[test]
-fn resolver_records_behavior_impl_method_body_locals() {
-    let program = parse_program(
-        r#"
-Json: behavior {
-    stringify: (Self) str
-}
-
-Point: { x: i32 }
-
-Point.implements(Json) {
-    stringify = (value: Point) str {
-        label = "point"
-        label
-    }
-}
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-    let value = table
-        .lookup_scoped(Namespace::Local, "value")
-        .expect("impl method parameter symbol");
-    let label = table
-        .lookup_scoped(Namespace::Local, "label")
-        .expect("impl method body local symbol");
-
-    assert_ne!(value.id, label.id);
-    assert_ne!(value.scope_id, label.scope_id);
-    assert_eq!(value.is_mutable, Some(false));
-    assert_eq!(label.is_mutable, Some(false));
-}
-
-#[test]
-fn resolver_accepts_behavior_requires_known_type_and_behavior() {
-    let program = parse_program(
-        r#"
-Json: behavior {
-    stringify: (Self) str
-}
-
-Point: { x: i32 }
-
-Point.requires(Json)
-"#,
-    );
-
-    Resolver::new()
-        .resolve_program(&program)
-        .expect("known requires assertion should resolve");
-}
-
-#[test]
-fn resolver_rejects_duplicate_behavior_required_edges() {
-    let program = parse_program(
-        r#"
-Json: behavior {
-    stringify: (Self) str
-}
-
-Point: { x: i32 }
-
-Point.requires(Json)
-Point.requires(Json)
-"#,
-    );
-
-    let err = Resolver::new()
-        .resolve_program(&program)
-        .expect_err("duplicate behavior requires should fail in resolver");
-
-    assert!(
-        err.iter()
-            .any(|d| d.message.contains("duplicate required behavior `Json`")),
-        "expected duplicate required behavior diagnostic, got {err:?}"
-    );
-}
-
-#[test]
-fn resolver_rejects_behavior_requires_unknown_symbols() {
-    let program = parse_program("Missing.requires(Json)");
-
-    let err = Resolver::new()
-        .resolve_program(&program)
-        .expect_err("unknown requires symbols should fail");
-    assert!(
-        err.iter()
-            .any(|d| d.message.contains("unknown type symbol 'Missing'")),
-        "expected unknown type diagnostic, got {err:?}"
-    );
-    assert!(
-        err.iter()
-            .any(|d| d.message.contains("unknown behavior symbol 'Json'")),
-        "expected unknown behavior diagnostic, got {err:?}"
-    );
-}
-
-#[test]
-fn resolver_accepts_behavior_extends_known_behaviors() {
-    let program = parse_program(
-        r#"
-Json: behavior {
-    stringify: (Self) str
-}
-
-PrettyJson: behavior {
-    pretty: (Self) str
-}
-
-PrettyJson.extends(Json)
-"#,
-    );
-
-    Resolver::new()
-        .resolve_program(&program)
-        .expect("known behavior inheritance should resolve");
-}
-
-#[test]
-fn resolver_rejects_duplicate_behavior_parent_edges() {
-    let program = parse_program(
-        r#"
-Json: behavior {
-    stringify: (Self) str
-}
-
-PrettyJson: behavior {
-    pretty: (Self) str
-}
-
-PrettyJson.extends(Json)
-PrettyJson.extends(Json)
-"#,
-    );
-
-    let err = Resolver::new()
-        .resolve_program(&program)
-        .expect_err("duplicate behavior inheritance should fail in resolver");
-
-    assert!(
-        err.iter()
-            .any(|d| d.message.contains("duplicate behavior parent `Json`")),
-        "expected duplicate behavior parent diagnostic, got {err:?}"
-    );
-}
-
-#[test]
-fn resolver_rejects_behavior_extends_unknown_symbols() {
-    let program = parse_program("PrettyJson.extends(Json)");
-
-    let err = Resolver::new()
-        .resolve_program(&program)
-        .expect_err("unknown behavior inheritance symbols should fail");
-    assert!(
-        err.iter()
-            .any(|d| d.message.contains("unknown behavior symbol 'PrettyJson'")),
-        "expected unknown child behavior diagnostic, got {err:?}"
-    );
-    assert!(
-        err.iter()
-            .any(|d| d.message.contains("unknown behavior symbol 'Json'")),
-        "expected unknown parent behavior diagnostic, got {err:?}"
-    );
-}
-
-#[test]
-fn resolver_records_value_symbol_parameter_counts() {
-    let program = parse_program(
-        r#"
-add = (a: i32, b: i32) i32 { a + b }
-Point: { x: i32 }
-Point.shift = (self: Point, dx: i32) Point { self }
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-
-    assert_eq!(
-        table
-            .lookup(Namespace::Value, "add")
-            .expect("function symbol")
-            .parameter_count,
-        Some(2)
-    );
-    assert_eq!(
-        table
-            .lookup(Namespace::Value, "Point.shift")
-            .expect("method symbol")
-            .parameter_count,
-        Some(2)
-    );
-}
-
-#[test]
-fn resolver_records_value_symbol_parameter_types() {
-    let program = parse_program(
-        r#"
-add = (a: i32, b: f64) f64 { b }
-Point: { x: i32 }
-Point.shift = (self: Point, dx: i32) Point { self }
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-
-    assert_eq!(
-        table
-            .lookup(Namespace::Value, "add")
-            .expect("function symbol")
-            .parameter_type_names
-            .as_deref(),
-        Some(&["i32".to_string(), "f64".to_string()][..])
-    );
-    assert_eq!(
-        table
-            .lookup(Namespace::Value, "Point.shift")
-            .expect("method symbol")
-            .parameter_type_names
-            .as_deref(),
-        Some(&["Point".to_string(), "i32".to_string()][..])
-    );
-}
-
-#[test]
-fn resolver_records_value_symbol_parameter_names() {
-    let program = parse_program(
-        r#"
-add = (a: i32, b: f64) f64 { b }
-Point: { x: i32 }
-Point.shift = (self: Point, dx: i32) Point { self }
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-
-    assert_eq!(
-        table
-            .lookup(Namespace::Value, "add")
-            .expect("function symbol")
-            .parameter_names
-            .as_deref(),
-        Some(&["a".to_string(), "b".to_string()][..])
-    );
-    assert_eq!(
-        table
-            .lookup(Namespace::Value, "Point.shift")
-            .expect("method symbol")
-            .parameter_names
-            .as_deref(),
-        Some(&["self".to_string(), "dx".to_string()][..])
-    );
-}
-
-#[test]
-fn resolver_records_value_symbol_return_types() {
-    let program = parse_program(
-        r#"
-main = () i32 { 0 }
-log = () { }
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-
-    assert_eq!(
-        table
-            .lookup(Namespace::Value, "main")
-            .expect("main symbol")
-            .return_type_name
-            .as_deref(),
-        Some("i32")
-    );
-    assert_eq!(
-        table
-            .lookup(Namespace::Value, "log")
-            .expect("log symbol")
-            .return_type_name
-            .as_deref(),
-        Some("void")
-    );
-}
-
-#[test]
-fn resolver_records_value_symbol_function_type_metadata() {
-    let program = parse_program(
-        r#"
-apply = (callback: (i32) i32, value: i32) (i32) i32 { callback }
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-    let apply = table
-        .lookup(Namespace::Value, "apply")
-        .expect("function symbol");
-
-    assert_eq!(
-        apply.parameter_type_names.as_deref(),
-        Some(&["(i32) i32".to_string(), "i32".to_string()][..])
-    );
-    assert_eq!(apply.return_type_name.as_deref(), Some("(i32) i32"));
-}
-
-#[test]
-fn resolver_records_value_symbol_generic_parameter_counts() {
-    let program = parse_program(
-        r#"
-identity<T> = (value: T) T { value }
-Point: { x: i32 }
-Point.wrap<T> = (self: Point, value: T) Point { self }
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-
-    assert_eq!(
-        table
-            .lookup(Namespace::Value, "identity")
-            .expect("function symbol")
-            .type_parameter_count,
-        Some(1)
-    );
-    assert_eq!(
-        table
-            .lookup(Namespace::Value, "identity")
-            .expect("function symbol")
-            .type_parameter_names
-            .as_ref()
-            .map(Vec::as_slice),
-        Some(&["T".to_string()][..])
-    );
-    assert_eq!(
-        table
-            .lookup(Namespace::Value, "Point.wrap")
-            .expect("method symbol")
-            .type_parameter_count,
-        Some(1)
-    );
-    assert_eq!(
-        table
-            .lookup(Namespace::Value, "Point.wrap")
-            .expect("method symbol")
-            .type_parameter_names
-            .as_ref()
-            .map(Vec::as_slice),
-        Some(&["T".to_string()][..])
-    );
-}
-
-#[test]
-fn resolver_records_value_symbol_generic_bounds() {
-    let program = parse_program(
-        r#"
-Json: behavior {
-    encode: (Self) str
-}
-encode<T: Json> = (value: T) str { "encoded" }
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-
-    assert_eq!(
-        table
-            .lookup(Namespace::Value, "encode")
-            .expect("function symbol")
-            .type_parameter_bounds
-            .as_ref()
-            .map(Vec::as_slice),
-        Some(&[("T".to_string(), "Json".to_string())][..])
-    );
-    assert_eq!(
-        table
-            .lookup(Namespace::Value, "encode")
-            .expect("function symbol")
-            .type_parameter_bound_refs
-            .as_ref()
-            .map(Vec::as_slice),
-        Some(
-            &[TypeParameterBoundRefMetadata {
-                type_parameter: "T".to_string(),
-                behavior: "Json".to_string(),
-                type_args: Vec::new(),
-            }][..]
-        )
-    );
-}
-
-#[test]
-fn resolver_records_type_and_behavior_generic_parameter_counts() {
-    let program = parse_program(
-        r#"
-Box<T>: { value: T }
-Option<T>: Some(T), None
-Serializable<T>: behavior {
-    encode: (T) str
-}
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-
-    assert_eq!(
-        table
-            .lookup(Namespace::Type, "Box")
-            .expect("struct symbol")
-            .type_parameter_count,
-        Some(1)
-    );
-    assert_eq!(
-        table
-            .lookup(Namespace::Type, "Box")
-            .expect("struct symbol")
-            .type_parameter_names
-            .as_ref()
-            .map(Vec::as_slice),
-        Some(&["T".to_string()][..])
-    );
-    assert_eq!(
-        table
-            .lookup(Namespace::Type, "Option")
-            .expect("enum symbol")
-            .type_parameter_count,
-        Some(1)
-    );
-    assert_eq!(
-        table
-            .lookup(Namespace::Type, "Option")
-            .expect("enum symbol")
-            .type_parameter_names
-            .as_ref()
-            .map(Vec::as_slice),
-        Some(&["T".to_string()][..])
-    );
-    assert_eq!(
-        table
-            .lookup(Namespace::Behavior, "Serializable")
-            .expect("behavior symbol")
-            .type_parameter_count,
-        Some(1)
-    );
-    assert_eq!(
-        table
-            .lookup(Namespace::Behavior, "Serializable")
-            .expect("behavior symbol")
-            .type_parameter_names
-            .as_ref()
-            .map(Vec::as_slice),
-        Some(&["T".to_string()][..])
-    );
-}
-
-#[test]
-fn resolver_rejects_duplicate_type_parameter_names() {
-    let program = parse_program(
-        r#"
-Box<T, T>: { value: T }
-Option<T, T>: Some(T), None
-Serializable<T, T>: behavior {
-    encode: (T) str
-}
-identity<T, T> = (value: T) T { value }
-"#,
-    );
-
-    let err = Resolver::new()
-        .resolve_program(&program)
-        .expect_err("duplicate type parameter names should fail in resolver");
-
-    let duplicate_count = err
-        .iter()
-        .filter(|d| d.message.contains("duplicate type parameter `T`"))
-        .count();
-    assert_eq!(
-        duplicate_count, 4,
-        "expected duplicate type parameter diagnostics for struct, enum, behavior, and function, got {err:?}"
-    );
-}
-
-#[test]
-fn resolver_records_type_and_behavior_generic_bounds() {
-    let program = parse_program(
-        r#"
-Json: behavior {
-    encode: (Self) str
-}
-Box<T: Json>: { value: T }
-Option<T: Json>: Some(T), None
-Serializable<T: Json>: behavior {
-    encode: (T) str
-}
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-
-    assert_eq!(
-        table
-            .lookup(Namespace::Type, "Box")
-            .expect("struct symbol")
-            .type_parameter_bounds
-            .as_ref()
-            .map(Vec::as_slice),
-        Some(&[("T".to_string(), "Json".to_string())][..])
-    );
-    assert_eq!(
-        table
-            .lookup(Namespace::Type, "Option")
-            .expect("enum symbol")
-            .type_parameter_bounds
-            .as_ref()
-            .map(Vec::as_slice),
-        Some(&[("T".to_string(), "Json".to_string())][..])
-    );
-    assert_eq!(
-        table
-            .lookup(Namespace::Behavior, "Serializable")
-            .expect("behavior symbol")
-            .type_parameter_bounds
-            .as_ref()
-            .map(Vec::as_slice),
-        Some(&[("T".to_string(), "Json".to_string())][..])
-    );
-}
-
-#[test]
-fn resolver_records_generic_behavior_bounds_with_type_args() {
-    let program = parse_program(
-        r#"
-Json<T>: behavior {
-    encode: (Self) T
-}
-Box<T: Json<T>>: { value: T }
-encode<T: Json<T>> = (value: T) T { value }
-Serializable<T: Json<T>>: behavior {
-    serialize: (Self) T
-}
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-
-    assert_eq!(
-        table
-            .lookup(Namespace::Type, "Box")
-            .expect("struct symbol")
-            .type_parameter_bounds
-            .as_ref()
-            .map(Vec::as_slice),
-        Some(&[("T".to_string(), "Json<T>".to_string())][..])
-    );
-    assert_eq!(
-        table
-            .lookup(Namespace::Value, "encode")
-            .expect("function symbol")
-            .type_parameter_bounds
-            .as_ref()
-            .map(Vec::as_slice),
-        Some(&[("T".to_string(), "Json<T>".to_string())][..])
-    );
-    assert_eq!(
-        table
-            .lookup(Namespace::Value, "encode")
-            .expect("function symbol")
-            .type_parameter_bound_refs
-            .as_ref()
-            .map(Vec::as_slice),
-        Some(
-            &[TypeParameterBoundRefMetadata {
-                type_parameter: "T".to_string(),
-                behavior: "Json".to_string(),
-                type_args: vec![zen::ast::AstType::Named("T".to_string())],
-            }][..]
-        )
-    );
-    assert_eq!(
-        table
-            .lookup(Namespace::Behavior, "Serializable")
-            .expect("behavior symbol")
-            .type_parameter_bounds
-            .as_ref()
-            .map(Vec::as_slice),
-        Some(&[("T".to_string(), "Json<T>".to_string())][..])
-    );
-}
-
-#[test]
-fn resolver_records_behavior_method_signatures() {
-    let program = parse_program(
-        r#"
-Serializable: behavior {
-    encode: (Self, i32) str
-    reset: () void
-}
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-
-    assert_eq!(
-        table
-            .lookup(Namespace::Behavior, "Serializable")
-            .expect("behavior symbol")
-            .behavior_method_signatures
-            .as_ref()
-            .map(Vec::as_slice),
-        Some(
-            &[
-                (
-                    "encode".to_string(),
-                    vec!["Self".to_string(), "i32".to_string()],
-                    "str".to_string()
-                ),
-                ("reset".to_string(), vec![], "void".to_string())
-            ][..]
-        )
-    );
-}
-
-#[test]
-fn resolver_rejects_duplicate_behavior_method_names() {
-    let program = parse_program(
-        r#"
-Serializable: behavior {
-    encode: (Self) str
-    encode: (Self, i32) str
-}
-"#,
-    );
-
-    let err = Resolver::new()
-        .resolve_program(&program)
-        .expect_err("duplicate behavior method names should fail in resolver");
-
-    assert!(
-        err.iter().any(|d| {
-            d.message
-                .contains("duplicate behavior method `encode` in `Serializable`")
-        }),
-        "expected duplicate behavior method diagnostic, got {err:?}"
-    );
-}
-
-#[test]
-fn resolver_rejects_duplicate_signature_parameter_names() {
-    let program = parse_program(
-        r#"
-Json: behavior {
-    encode: (value: Self, value: Self) str
-}
-"#,
-    );
-
-    let err = Resolver::new()
-        .resolve_program(&program)
-        .expect_err("duplicate behavior method parameter names should fail in resolver");
-
-    assert!(
-        err.iter()
-            .any(|d| d.message.contains("duplicate parameter `value`")),
-        "expected duplicate behavior method parameter diagnostic, got {err:?}"
-    );
-}
-
-#[test]
-fn resolver_records_behavior_function_type_method_signatures() {
-    let program = parse_program(
-        r#"
-Mapper: behavior {
-    map: (Self, (i32) i32) (i32) i32
-}
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-
-    assert_eq!(
-        table
-            .lookup(Namespace::Behavior, "Mapper")
-            .expect("behavior symbol")
-            .behavior_method_signatures
-            .as_ref()
-            .map(Vec::as_slice),
-        Some(
-            &[(
-                "map".to_string(),
-                vec!["Self".to_string(), "(i32) i32".to_string()],
-                "(i32) i32".to_string()
-            )][..]
-        )
-    );
-    assert_eq!(
-        table
-            .lookup(Namespace::Behavior, "Mapper")
-            .expect("behavior symbol")
-            .behavior_method_types
-            .as_ref()
-            .map(Vec::as_slice),
-        Some(
-            &[zen::resolver::BehaviorMethodTypeMetadata {
-                name: "map".to_string(),
-                parameter_names: vec!["__arg0".to_string(), "__arg1".to_string()],
-                parameter_types: vec![
-                    zen::ast::AstType::SelfType,
-                    zen::ast::AstType::Function {
-                        params: vec![zen::ast::AstType::I32],
-                        ret: Box::new(zen::ast::AstType::I32),
-                    },
-                ],
-                return_type: zen::ast::AstType::Function {
-                    params: vec![zen::ast::AstType::I32],
-                    ret: Box::new(zen::ast::AstType::I32),
-                },
-            }][..]
-        )
-    );
-}
-
-#[test]
-fn resolver_records_generic_behavior_method_signatures() {
-    let program = parse_program(
-        r#"
-Json<T>: behavior {
-    encode: (Self) T
-}
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-
-    assert_eq!(
-        table
-            .lookup(Namespace::Behavior, "Json")
-            .expect("behavior symbol")
-            .behavior_method_signatures
-            .as_ref()
-            .map(Vec::as_slice),
-        Some(
-            &[(
-                "encode".to_string(),
-                vec!["Self".to_string()],
-                "T".to_string()
-            )][..]
-        )
-    );
-}
-
-#[test]
-fn resolver_records_generic_behavior_function_type_method_signatures() {
-    let program = parse_program(
-        r#"
-Mapper<T>: behavior {
-    map: (Self, (T) T) (T) T
-}
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-
-    assert_eq!(
-        table
-            .lookup(Namespace::Behavior, "Mapper")
-            .expect("behavior symbol")
-            .behavior_method_signatures
-            .as_ref()
-            .map(Vec::as_slice),
-        Some(
-            &[(
-                "map".to_string(),
-                vec!["Self".to_string(), "(T) T".to_string()],
-                "(T) T".to_string()
-            )][..]
-        )
-    );
-}
-
-#[test]
-fn resolver_records_behavior_default_method_body_locals() {
-    let program = parse_program(
-        r#"
-Json: behavior {
-    stringify: (Self) str {
-        label = "json"
-        label
-    }
-}
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-    let label = table
-        .lookup_scoped(Namespace::Local, "label")
-        .expect("behavior default body local symbol");
-
-    assert_eq!(label.is_mutable, Some(false));
-    assert!(label.scope_id > 0);
-}
-
-#[test]
-fn resolver_records_behavior_parent_names() {
-    let program = parse_program(
-        r#"
-Json: behavior {
-    encode: (Self) str
-}
-PrettyJson: behavior {
-    pretty: (Self) str
-}
-
-PrettyJson.extends(Json)
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-
-    assert_eq!(
-        table
-            .lookup(Namespace::Behavior, "PrettyJson")
-            .expect("behavior symbol")
-            .behavior_parent_names
-            .as_ref()
-            .map(Vec::as_slice),
-        Some(&["Json".to_string()][..])
-    );
-}
-
-#[test]
-fn resolver_records_behavior_impl_and_requires_names() {
-    let program = parse_program(
-        r#"
-Json<T>: behavior {
-    encode: (Self) T
-}
-
-Point: { x: i32 }
-
-Point.implements(Json<str>) {
-    encode = (value: Point) str { "point" }
-}
-
-Point.requires(Json<str>)
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-    let point = table.lookup(Namespace::Type, "Point").expect("Point type");
-
-    assert_eq!(
-        point.behavior_impl_names.as_ref().map(Vec::as_slice),
-        Some(&["Json<str>".to_string()][..])
-    );
-    assert_eq!(
-        point.behavior_impl_refs.as_ref().map(Vec::as_slice),
-        Some(
-            &[zen::resolver::BehaviorRefMetadata {
-                name: "Json".to_string(),
-                type_args: vec![zen::ast::AstType::Str],
-            }][..]
-        )
-    );
-    assert_eq!(
-        point.behavior_required_names.as_ref().map(Vec::as_slice),
-        Some(&["Json<str>".to_string()][..])
-    );
-    assert_eq!(
-        point.behavior_required_refs.as_ref().map(Vec::as_slice),
-        Some(
-            &[zen::resolver::BehaviorRefMetadata {
-                name: "Json".to_string(),
-                type_args: vec![zen::ast::AstType::Str],
-            }][..]
-        )
-    );
-}
-
-#[test]
-fn resolver_records_generic_behavior_parent_names() {
-    let program = parse_program(
-        r#"
-Json<T>: behavior {
-    encode: (Self) T
-}
-PrettyJson: behavior {
-    pretty: (Self) str
-}
-
-PrettyJson.extends(Json<str>)
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-
-    assert_eq!(
-        table
-            .lookup(Namespace::Behavior, "PrettyJson")
-            .expect("behavior symbol")
-            .behavior_parent_names
-            .as_ref()
-            .map(Vec::as_slice),
-        Some(&["Json<str>".to_string()][..])
-    );
-    assert_eq!(
-        table
-            .lookup(Namespace::Behavior, "PrettyJson")
-            .expect("behavior symbol")
-            .behavior_parent_refs
-            .as_ref()
-            .map(Vec::as_slice),
-        Some(
-            &[zen::resolver::BehaviorRefMetadata {
-                name: "Json".to_string(),
-                type_args: vec![zen::ast::AstType::Str],
-            }][..]
-        )
-    );
-}
-
-#[test]
-fn resolver_accepts_behavior_parent_type_args_from_child_type_params() {
-    let program = parse_program(
-        r#"
-Json<T>: behavior {
-    encode: (Self) T
-}
-Serializable<T: Json<T>>: behavior {
-    serialize: (Self) T
-}
-Pretty<T: Json<T>>: behavior {
-    pretty: (Self) T
-}
-
-Pretty.extends(Serializable<T>)
-"#,
-    );
-
-    let table = Resolver::new()
-        .resolve_program(&program)
-        .expect("generic behavior parent should accept child type parameter args");
-
-    assert_eq!(
-        table
-            .lookup(Namespace::Behavior, "Pretty")
-            .expect("behavior symbol")
-            .behavior_parent_refs
-            .as_ref()
-            .map(Vec::as_slice),
-        Some(
-            &[zen::resolver::BehaviorRefMetadata {
-                name: "Serializable".to_string(),
-                type_args: vec![zen::ast::AstType::Named("T".to_string())],
-            }][..]
-        )
-    );
-}
-
-#[test]
-fn resolver_rejects_behavior_parent_type_args_outside_child_type_params() {
-    let program = parse_program(
-        r#"
-Json<T>: behavior {
-    encode: (Self) T
-}
-Serializable<T: Json<T>>: behavior {
-    serialize: (Self) T
-}
-Pretty<T: Json<T>>: behavior {
-    pretty: (Self) T
-}
-
-Pretty.extends(Serializable<U>)
-"#,
-    );
-
-    let err = Resolver::new()
-        .resolve_program(&program)
-        .expect_err("unknown parent type arg should fail in resolver");
-
-    assert!(
-        err.iter()
-            .any(|d| d.message.contains("unknown type symbol 'U'")),
-        "expected unknown type parameter diagnostic, got {err:?}"
-    );
-}
-
-#[test]
-fn resolver_records_struct_field_counts() {
-    let program = parse_program(
-        r#"
-Point: { x: i32, y: i32 }
-Empty: { }
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-
-    assert_eq!(
-        table
-            .lookup(Namespace::Type, "Point")
-            .expect("Point symbol")
-            .field_count,
-        Some(2)
-    );
-    assert_eq!(
-        table
-            .lookup(Namespace::Type, "Empty")
-            .expect("Empty symbol")
-            .field_count,
-        Some(0)
-    );
-}
-
-#[test]
-fn resolver_records_struct_field_types() {
-    let program = parse_program(
-        r#"
-Point: { x: i32, y: f64 }
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-
-    assert_eq!(
-        table
-            .lookup(Namespace::Type, "Point")
-            .expect("Point symbol")
-            .field_type_names
-            .as_ref()
-            .map(Vec::as_slice),
-        Some(
-            &[
-                ("x".to_string(), "i32".to_string()),
-                ("y".to_string(), "f64".to_string())
-            ][..]
-        )
-    );
-}
-
-#[test]
-fn resolver_rejects_duplicate_struct_field_names() {
-    let program = parse_program(
-        r#"
-Point: { x: i32, x: i64 }
-"#,
-    );
-
-    let err = Resolver::new()
-        .resolve_program(&program)
-        .expect_err("duplicate struct field names should fail in resolver");
-
-    assert!(
-        err.iter()
-            .any(|d| d.message.contains("duplicate field `x` for struct `Point`")),
-        "expected duplicate struct field diagnostic, got {err:?}"
-    );
-}
-
-#[test]
-fn resolver_records_struct_function_type_fields() {
-    let program = parse_program(
-        r#"
-Pipeline: { callback: (i32) i32 }
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-
-    assert_eq!(
-        table
-            .lookup(Namespace::Type, "Pipeline")
-            .expect("Pipeline symbol")
-            .field_type_names
-            .as_ref()
-            .map(Vec::as_slice),
-        Some(&[("callback".to_string(), "(i32) i32".to_string())][..])
-    );
-    assert_eq!(
-        table
-            .lookup(Namespace::Type, "Pipeline")
-            .expect("Pipeline symbol")
-            .field_types
-            .as_ref()
-            .map(Vec::as_slice),
-        Some(
-            &[(
-                "callback".to_string(),
-                zen::ast::AstType::Function {
-                    params: vec![zen::ast::AstType::I32],
-                    ret: Box::new(zen::ast::AstType::I32),
-                }
-            )][..]
-        )
-    );
-}
-
-#[test]
-fn resolver_records_generic_struct_field_types() {
-    let program = parse_program(
-        r#"
-Box<T>: { value: T }
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-
-    assert_eq!(
-        table
-            .lookup(Namespace::Type, "Box")
-            .expect("Box symbol")
-            .field_type_names
-            .as_ref()
-            .map(Vec::as_slice),
-        Some(&[("value".to_string(), "T".to_string())][..])
-    );
-}
-
-#[test]
-fn resolver_records_struct_field_default_locals() {
-    let program = parse_program(
-        r#"
-Point: {
-    x: i32 = {
-        value = 1
-        value
-    }
-}
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-    let value = table
-        .lookup_scoped(Namespace::Local, "value")
-        .expect("struct field default local symbol");
-
-    assert_eq!(value.is_mutable, Some(false));
-    assert!(value.scope_id > 0);
-}
-
-#[test]
-fn resolver_rejects_duplicate_struct_literal_fields() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-main = () i32 {
-    point = Point { x: 1, x: 2 }
-    0
-}
-"#,
-    );
-
-    let err = Resolver::new()
-        .resolve_program(&program)
-        .expect_err("duplicate struct literal field should fail in resolver");
-
-    assert!(
-        err.iter()
-            .any(|d| d.message.contains("duplicate field `x` for struct `Point`")),
-        "expected duplicate struct literal field diagnostic, got {err:?}"
-    );
-}
-
-#[test]
-fn resolver_rejects_unknown_struct_literal_fields() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-main = () i32 {
-    point = Point { x: 1, y: 2 }
-    0
-}
-"#,
-    );
-
-    let err = Resolver::new()
-        .resolve_program(&program)
-        .expect_err("unknown struct literal field should fail in resolver");
-
-    assert!(
-        err.iter()
-            .any(|d| d.message.contains("unknown field `y` for struct `Point`")),
-        "expected unknown struct literal field diagnostic, got {err:?}"
-    );
-}
-
-#[test]
-fn resolver_rejects_missing_struct_literal_fields() {
-    let program = parse_program(
-        r#"
-Point: { x: i32, y: i32 }
-
-main = () i32 {
-    point = Point { x: 1 }
-    0
-}
-"#,
-    );
-
-    let err = Resolver::new()
-        .resolve_program(&program)
-        .expect_err("missing struct literal field should fail in resolver");
-
-    assert!(
-        err.iter()
-            .any(|d| d.message.contains("missing field `y` for struct `Point`")),
-        "expected missing struct literal field diagnostic, got {err:?}"
-    );
-}
-
-#[test]
-fn resolver_rejects_unknown_struct_literal_types() {
-    let program = parse_program(
-        r#"
-main = () i32 {
-    point = Point { x: 1 }
-    0
-}
-"#,
-    );
-
-    let err = Resolver::new()
-        .resolve_program(&program)
-        .expect_err("unknown struct literal type should fail in resolver");
-
-    assert!(
-        err.iter()
-            .any(|d| d.message.contains("unknown type symbol 'Point'")),
-        "expected unknown struct literal type diagnostic, got {err:?}"
-    );
-}
-
-#[test]
-fn resolver_records_enum_variant_payload_counts() {
-    let program = parse_program(
-        r#"
-Option: Some(i32), None
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-
-    assert_eq!(
-        table
-            .lookup(Namespace::Variant, "Some")
-            .expect("Some variant symbol")
-            .variant_payload_count,
-        Some(1)
-    );
-    assert_eq!(
-        table
-            .lookup(Namespace::Variant, "None")
-            .expect("None variant symbol")
-            .variant_payload_count,
-        Some(0)
-    );
-}
-
-#[test]
-fn resolver_records_enum_variant_names() {
-    let program = parse_program(
-        r#"
-Option: Some(i32), None
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-
-    assert_eq!(
-        table
-            .lookup(Namespace::Type, "Option")
-            .expect("Option type symbol")
-            .variant_names
-            .as_ref()
-            .map(Vec::as_slice),
-        Some(&["Some".to_string(), "None".to_string()][..])
-    );
-}
-
-#[test]
-fn resolver_allows_same_variant_names_in_different_enums() {
-    let program = parse_program(
-        r#"
-Option:
-    None,
-    Some(i32)
-
-Maybe:
-    None,
-    Some(bool)
-"#,
-    );
-
-    let table = Resolver::new()
-        .resolve_program(&program)
-        .expect("variant names should be scoped to their owner enum");
-
-    assert_eq!(
-        table
-            .symbols()
-            .iter()
-            .filter(|symbol| symbol.namespace == Namespace::Variant && symbol.name == "None")
-            .count(),
-        2
-    );
-    assert_eq!(
-        table
-            .symbols()
-            .iter()
-            .filter(|symbol| symbol.namespace == Namespace::Variant && symbol.name == "Some")
-            .count(),
-        2
-    );
-}
-
-#[test]
-fn resolver_rejects_duplicate_variant_names_in_same_enum() {
-    let program = parse_program(
-        r#"
-Option:
-    None,
-    None
-"#,
-    );
-
-    let err = Resolver::new()
-        .resolve_program(&program)
-        .expect_err("duplicate variant names in one enum should be rejected");
-
-    assert!(
-        err.iter()
-            .any(|d| d.message.contains("duplicate variant symbol 'None'")),
-        "expected duplicate variant diagnostic, got {err:?}"
-    );
-}
-
-#[test]
-fn resolver_records_enum_variant_owner_names() {
-    let program = parse_program(
-        r#"
-Option: Some(i32), None
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-
-    assert_eq!(
-        table
-            .lookup(Namespace::Variant, "Some")
-            .expect("Some variant symbol")
-            .variant_owner_name
-            .as_deref(),
-        Some("Option")
-    );
-    assert_eq!(
-        table
-            .lookup(Namespace::Variant, "None")
-            .expect("None variant symbol")
-            .variant_owner_name
-            .as_deref(),
-        Some("Option")
-    );
-}
-
-#[test]
-fn resolver_records_enum_variant_payload_types() {
-    let program = parse_program(
-        r#"
-Option: Some(i32), None
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-
-    assert_eq!(
-        table
-            .lookup(Namespace::Variant, "Some")
-            .expect("Some variant symbol")
-            .variant_payload_type_name
-            .as_deref(),
-        Some("i32")
-    );
-    assert_eq!(
-        table
-            .lookup(Namespace::Variant, "None")
-            .expect("None variant symbol")
-            .variant_payload_type_name
-            .as_deref(),
-        None
-    );
-}
-
-#[test]
-fn resolver_records_generic_enum_variant_payload_types() {
-    let program = parse_program(
-        r#"
-Result<T, E>: Ok(T), Err(E)
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-
-    assert_eq!(
-        table
-            .lookup(Namespace::Variant, "Ok")
-            .expect("Ok variant symbol")
-            .variant_payload_type_name
-            .as_deref(),
-        Some("T")
-    );
-    assert_eq!(
-        table
-            .lookup(Namespace::Variant, "Err")
-            .expect("Err variant symbol")
-            .variant_payload_type_name
-            .as_deref(),
-        Some("E")
-    );
-}
-
-#[test]
-fn resolver_records_enum_function_type_payloads() {
-    let program = parse_program(
-        r#"
-Callback: Wrap((i32) i32), None
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-
-    assert_eq!(
-        table
-            .lookup(Namespace::Variant, "Wrap")
-            .expect("Wrap variant symbol")
-            .variant_payload_type_name
-            .as_deref(),
-        Some("(i32) i32")
-    );
-    assert_eq!(
-        table
-            .lookup(Namespace::Variant, "Wrap")
-            .expect("Wrap variant symbol")
-            .variant_payload_type
-            .as_ref(),
-        Some(&zen::ast::AstType::Function {
-            params: vec![zen::ast::AstType::I32],
-            ret: Box::new(zen::ast::AstType::I32),
-        })
-    );
-}
-
-#[test]
-fn resolver_records_generic_enum_function_type_payloads() {
-    let program = parse_program(
-        r#"
-Callback<T>: Wrap((T) T), None
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-
-    assert_eq!(
-        table
-            .lookup(Namespace::Variant, "Wrap")
-            .expect("Wrap variant symbol")
-            .variant_payload_type_name
-            .as_deref(),
-        Some("(T) T")
-    );
-}
-
-#[test]
-fn resolver_rejects_unknown_unqualified_function_calls() {
-    let program = parse_program(
-        r#"
-known = () i32 { 1 }
-main = () i32 { missing() }
-"#,
-    );
-
-    let err = Resolver::new()
-        .resolve_program(&program)
-        .expect_err("unknown function call should fail");
-
-    assert!(
-        err.iter()
-            .any(|d| d.message.contains("unknown value symbol 'missing'")),
-        "expected missing function diagnostic, got {err:?}"
-    );
-}
-
-#[test]
-fn resolver_records_parameter_and_local_symbols() {
-    let program = parse_program(
-        r#"
-main = (mut input: i32) i32 {
-    value ::= input
-    frozen = value
-    frozen
-}
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-    let input = table
-        .lookup_scoped(Namespace::Local, "input")
-        .expect("parameter symbol");
-    let value = table
-        .lookup_scoped(Namespace::Local, "value")
-        .expect("local symbol");
-    let frozen = table
-        .lookup_scoped(Namespace::Local, "frozen")
-        .expect("immutable local symbol");
-
-    assert_ne!(input.id, value.id);
-    assert_ne!(input.scope_id, value.scope_id);
-    assert_eq!(input.is_mutable, Some(true));
-    assert_eq!(value.is_mutable, Some(true));
-    assert_eq!(frozen.is_mutable, Some(false));
-}
-
-#[test]
-fn resolver_records_top_level_expr_locals() {
-    let program = parse_program(
-        r#"
-value := 1
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-    let value = table
-        .lookup_scoped(Namespace::Local, "value")
-        .expect("top-level expr local symbol");
-
-    assert_eq!(value.is_mutable, Some(false));
-    assert!(value.scope_id > 0);
-}
-
-#[test]
-fn resolver_rejects_unknown_enum_variant_expressions() {
-    let program = parse_program(
-        r#"
-Status: Ok, Err
-
-main = () i32 {
-    value = Status.Pending
-    0
-}
-"#,
-    );
-
-    let err = Resolver::new()
-        .resolve_program(&program)
-        .expect_err("unknown enum variant expression should fail in resolver");
-
-    assert!(
-        err.iter()
-            .any(|d| d.message.contains("enum `Status` has no variant `Pending`")),
-        "expected unknown enum variant diagnostic, got {err:?}"
-    );
-}
-
-#[test]
-fn resolver_rejects_missing_enum_variant_payload_expressions() {
-    let program = parse_program(
-        r#"
-Maybe: Some(i32), None
-
-main = () i32 {
-    value = Maybe.Some
-    0
-}
-"#,
-    );
-
-    let err = Resolver::new()
-        .resolve_program(&program)
-        .expect_err("missing enum variant payload expression should fail in resolver");
-
-    assert!(
-        err.iter().any(|d| d
-            .message
-            .contains("enum variant `Maybe.Some` requires a payload")),
-        "expected missing enum variant payload diagnostic, got {err:?}"
-    );
-}
-
-#[test]
-fn resolver_rejects_unexpected_enum_variant_payload_expressions() {
-    let program = parse_program(
-        r#"
-Maybe: Some(i32), None
-
-main = () i32 {
-    value = Maybe.None(1)
-    0
-}
-"#,
-    );
-
-    let err = Resolver::new()
-        .resolve_program(&program)
-        .expect_err("unexpected enum variant payload expression should fail in resolver");
-
-    assert!(
-        err.iter().any(|d| d
-            .message
-            .contains("enum variant `Maybe.None` does not accept a payload")),
-        "expected unexpected enum variant payload diagnostic, got {err:?}"
-    );
-}
-
-#[test]
-fn resolver_records_closure_locals() {
-    let program = parse_program(
-        r#"
-main = () i32 {
-    mapper = (input: i32) i32 {
-        inner = input
-        inner
-    }
-    0
-}
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-    let mapper = table
-        .lookup_scoped(Namespace::Local, "mapper")
-        .expect("closure binding local symbol");
-    let input = table
-        .lookup_scoped(Namespace::Local, "input")
-        .expect("closure parameter local symbol");
-    let inner = table
-        .lookup_scoped(Namespace::Local, "inner")
-        .expect("closure body local symbol");
-
-    assert_ne!(mapper.scope_id, input.scope_id);
-    assert_ne!(input.scope_id, inner.scope_id);
-    assert!(inner.scope_id > input.scope_id);
-    assert_eq!(mapper.is_mutable, Some(false));
-    assert_eq!(input.is_mutable, Some(false));
-    assert_eq!(inner.is_mutable, Some(false));
-}
-
-#[test]
-fn resolver_records_mutable_closure_parameter_locals() {
-    let program = parse_program(
-        r#"
-main = () i32 {
-    mapper = (mut input: i32) i32 {
-        input = input + 1
-        input
-    }
-    0
-}
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-    let input = table
-        .lookup_scoped(Namespace::Local, "input")
-        .expect("closure parameter local symbol");
-
-    assert_eq!(input.is_mutable, Some(true));
-}
-
-#[test]
-fn resolver_records_pattern_locals() {
-    let program = parse_program(
-        r#"
-Option:
-    None,
-    Some(i32)
-
-main = (value: Option) i32 {
-    value ?
-        | Some(inner) { inner }
-        | None { 0 }
-}
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-    let value = table
-        .lookup_scoped(Namespace::Local, "value")
-        .expect("parameter local symbol");
-    let inner = table
-        .lookup_scoped(Namespace::Local, "inner")
-        .expect("pattern local symbol");
-
-    assert_ne!(value.scope_id, inner.scope_id);
-    assert_eq!(inner.is_mutable, Some(false));
-    assert!(inner.scope_id > value.scope_id);
-}
-
-#[test]
-fn resolver_records_same_name_locals_in_distinct_scopes() {
-    let program = parse_program(
-        r#"
-main = () i32 {
-    value := 1
-    {
-        value := 2
-        value
-    }
-    value
-}
-"#,
-    );
-
-    let table = Resolver::new().resolve_program(&program).expect("resolve");
-    let values: Vec<_> = table
-        .symbols()
-        .iter()
-        .filter(|symbol| symbol.namespace == Namespace::Local && symbol.name == "value")
-        .collect();
-
-    assert_eq!(values.len(), 2);
-    assert_ne!(values[0].id, values[1].id);
-    assert_ne!(values[0].scope_id, values[1].scope_id);
-    assert!(values.iter().all(|symbol| symbol.is_mutable == Some(false)));
-}
-
-#[test]
-fn resolver_rejects_duplicate_bindings_in_same_scope() {
-    let program = parse_program(
-        r#"
-main = (input: i32, input: i32) i32 {
-    value = 1
-    value = 2
-    value
-}
-"#,
-    );
-
-    let err = Resolver::new()
-        .resolve_program(&program)
-        .expect_err("duplicate locals should fail");
-
-    assert!(
-        err.iter()
-            .any(|d| d.message.contains("duplicate local symbol 'input'")),
-        "expected duplicate parameter diagnostic, got {err:?}"
-    );
-    assert!(
-        err.iter()
-            .any(|d| d.message.contains("duplicate local symbol 'value'")),
-        "expected duplicate local diagnostic, got {err:?}"
-    );
-}
-
-#[test]
-fn resolver_rejects_unknown_local_identifier_references() {
-    let program = parse_program(
-        r#"
-main = () i32 {
-    missing_local
-}
-"#,
-    );
-
-    let err = Resolver::new()
-        .resolve_program(&program)
-        .expect_err("unknown local identifier should fail");
-
-    assert!(
-        err.iter()
-            .any(|d| d.message.contains("unknown value symbol 'missing_local'")),
-        "expected missing local diagnostic, got {err:?}"
-    );
-}
+#[path = "resolver_phase2/behavior_relations.rs"]
+mod behavior_relations;
+#[path = "resolver_phase2/core_symbols.rs"]
+mod core_symbols;
+#[path = "resolver_phase2/enum_metadata.rs"]
+mod enum_metadata;
+#[path = "resolver_phase2/expr_locals.rs"]
+mod expr_locals;
+#[path = "resolver_phase2/generic_behavior_metadata.rs"]
+mod generic_behavior_metadata;
+#[path = "resolver_phase2/impls.rs"]
+mod impls;
+#[path = "resolver_phase2/struct_metadata.rs"]
+mod struct_metadata;
+#[path = "resolver_phase2/value_metadata.rs"]
+mod value_metadata;
